feat(seahaven-package): allow only one service per package
- Moved the `Manifest`, `PackageMeta`, `Service`, and `InitContainer` structs to a new `model.rs` file for better organization.
- Updated the `Manifest` struct to allow only one service instead of multiple, changing the corresponding serialization and deserialization logic.
- Adjusted TOML test data to reflect the new service structure, ensuring compatibility with the updated manifest format.
- Enhanced error messages for clarity when no service or init containers are defined.

This refactor improves the clarity and maintainability of the manifest handling in the Seahaven package.
