+++ seahaven-package/testdata/src/kitchen_sink.rs
@@ -12,12 +12,12 @@
 #[doc = r##"readme = "README.md""##]
 #[doc = r##""##]
 #[doc = r##""##]
-#[doc = r##"# Service target settings. Zero or more are allowed per package."##]
-#[doc = r##"[[service]]"##]
+#[doc = r##"# Service target settings. Zero or one are allowed per package."##]
+#[doc = r##"[service]"##]
 #[doc = r##"name = "chain""##]
 #[doc = r##"image = "ghcr.io/foundry-rs/foundry:latest""##]
 #[doc = r##""##]
-#[doc = r##"[[service.defaults]]"##]
+#[doc = r##"[service.defaults]"##]
 #[doc = r##"command = "anvil --host=0.0.0.0 --chain-id=${CHAIN_ID} --base-fee=0""##]
 #[doc = r##"ports = [ "${CHAIN_RPC}:8545" ]"##]
 #[doc = r##"healthcheck = { interval = "1s", retries = 10, test = "cast block" }"##]
@@ -67,12 +67,12 @@ pub const KITCHEN_SINK: &str = indoc::indoc! { r###"
   readme = "README.md"
 
 
-  # Service target settings. Zero or more are allowed per package.
-  [[service]]
+  # Service target settings. Zero or one are allowed per package.
+  [service]
   name = "chain"
   image = "ghcr.io/foundry-rs/foundry:latest"
 
-  [[service.defaults]]
+  [service.defaults]
   command = "anvil --host=0.0.0.0 --chain-id=${CHAIN_ID} --base-fee=0"
   ports = [ "${CHAIN_RPC}:8545" ]
   healthcheck = { interval = "1s", retries = 10, test = "cast block" }
